feat: update WalletDropdown import paths and add new wallet components
