+++ packages/ui/src/components/wallet-dropdown/wallet-switch.tsx
@@ -10,16 +10,16 @@ import {
   DropdownMenuRadioGroup,
   DropdownMenuRadioItem,
   DropdownMenuSeparator,
-} from "./dropdown-menu";
+} from "../dropdown-menu";
 
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 import {
   Accordion,
   AccordionContent,
   AccordionItem,
   AccordionTrigger,
-} from "./accordion";
-import { ScrollArea } from "./scroll-area";
+} from "../accordion";
+import { ScrollArea } from "../scroll-area";
 import { InjectedAccountWithMeta } from "./wallet-dropdown";
 
 interface WalletSwitchProps {
